fixed wrong prototype

+++ ags/audio/ags_note.c
@@ -1654,7 +1654,7 @@ ags_note_new()
 AgsNote*
 ags_note_new_with_offset(guint x0, guint x1,
 			 guint y,
-			 gdouble stream_delay, guint stream_attack)
+			 gdouble stream_delay, gdouble stream_attack)
 {
   AgsNote *note;
